Introduce extra rule pruning in RuleCache (#5046)
## What is the goal of this PR?

Improve and test query execution time scaling with number of rules.

## What are the changes implemented in this PR?

- introduced rule pruning based on type instance absence in the rule graph
- added conjunctive query shortcircuiting based on the type instance absence
- added a scalable non-trivial test stressing the rule application aspect
- refactored `RuleApplicabilityIT`

## Future work
- the prunable types should be stored/extracted from db statistics layer
