+++ server/src/server/session/cache/RuleCache.java
@@ -25,13 +25,14 @@ import grakn.core.concept.type.SchemaConcept;
 import grakn.core.concept.type.Type;
 import grakn.core.server.kb.Schema;
 import grakn.core.server.session.TransactionOLTP;
-
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.Map;
 import java.util.Set;
 import java.util.function.Supplier;
-import java.util.stream.Collectors;
 import java.util.stream.Stream;
+
+import static java.util.stream.Collectors.toSet;
 
 /**
  * Caches rules applicable to schema concepts and their conversion to InferenceRule object (parsing is expensive when large number of rules present).
@@ -42,6 +43,12 @@ public class RuleCache {
     private final HashMultimap<Type, Rule> ruleMap = HashMultimap.create();
     private final Map<Rule, Object> ruleConversionMap = new HashMap<>();
     private final TransactionOLTP tx;
+
+    //TODO: these should be eventually stored together with statistics
+    private Set<Type> absentTypes = new HashSet<>();
+    private Set<Type> checkedTypes = new HashSet<>();
+    private Set<Rule> fruitlessRules = new HashSet<>();
+    private Set<Rule> checkedRules = new HashSet<>();
 
     public RuleCache(TransactionOLTP tx) {
         this.tx = tx;
@@ -76,9 +83,9 @@ public class RuleCache {
      * @return relevant part (direct only or subs) of the type hierarchy of a type
      */
     private Set<Type> getTypes(Type type, boolean direct) {
-        Set<Type> types = direct ? Sets.newHashSet(type) : type.subs().collect(Collectors.toSet());
+        Set<Type> types = direct ? Sets.newHashSet(type) : type.subs().collect(toSet());
         return type.isImplicit() ?
-                types.stream().flatMap(t -> Stream.of(t, tx.getType(Schema.ImplicitType.explicitLabel(t.label())))).collect(Collectors.toSet()) :
+                types.stream().flatMap(t -> Stream.of(t, tx.getType(Schema.ImplicitType.explicitLabel(t.label())))).collect(toSet()) :
                 types;
     }
 
@@ -88,6 +95,23 @@ public class RuleCache {
      */
     public Stream<Rule> getRulesWithType(Type type) {
         return getRulesWithType(type, false);
+    }
+
+    /**
+     * @param types to check
+     * @return true if any of the provided types is absent - doesn't have instances
+     */
+    public boolean absentTypes(Set<Type> types) {
+        return types.stream().anyMatch(t -> !typeHasInstances(t));
+    }
+
+    /**
+     * acknowledge addition of an instance of a specific type
+     * @param type to be acked
+     */
+    public void ackTypeInstance(Type type){
+        checkedTypes.add(type);
+        absentTypes.remove(type);
     }
 
     /**
@@ -103,7 +127,36 @@ public class RuleCache {
 
         return getTypes(type, direct).stream()
                 .flatMap(SchemaConcept::thenRules)
+                .filter(this::checkRule)
                 .peek(rule -> ruleMap.put(type, rule));
+    }
+
+    private boolean typeHasInstances(Type type){
+        if (checkedTypes.contains(type)) return !absentTypes.contains(type);
+        checkedTypes.add(type);
+        boolean instancePresent = type.instances().findFirst().isPresent()
+                || type.thenRules().anyMatch(this::checkRule);
+        if (!instancePresent){
+            absentTypes.add(type);
+            type.whenRules().forEach(r -> fruitlessRules.add(r));
+        }
+        return instancePresent;
+    }
+
+    /**
+     *
+     * @param rule to be checked for matchability
+     * @return true if rule is matchable (can provide answers)
+     */
+    private boolean checkRule(Rule rule){
+        if (fruitlessRules.contains(rule)) return false;
+        if (checkedRules.contains(rule)) return true;
+        checkedRules.add(rule);
+        return rule.whenTypes()
+                //.filter(t -> !checkedTypes.contains(t))
+                .allMatch(this::typeHasInstances);
+                //.peek(t ->
+                //.findFirst().isPresent();
     }
 
     /**
@@ -127,5 +180,9 @@ public class RuleCache {
     public void clear() {
         ruleMap.clear();
         ruleConversionMap.clear();
+        absentTypes.clear();
+        checkedTypes.clear();
+        checkedRules.clear();
+        fruitlessRules.clear();
     }
 }
