+++ test-integration/graql/reasoner/query/NegationIT.java
@@ -375,7 +375,7 @@ public class NegationIT {
             EntityType specificType = tx.getEntityType(specificTypeLabel);
 
             List<ConceptMap> fullAnswers = tx.execute(Graql.parse("match $x has attribute $r;get;").asGet());
-            List<ConceptMap> answersWithoutSpecifcTypeAndValue = tx.execute(Graql.<GraqlGet>parse(
+            List<ConceptMap> answersWithoutSpecificTypeAndValue = tx.execute(Graql.<GraqlGet>parse(
                     "match " +
                             "$x has attribute $r;" +
                             "not {$x isa " + specificTypeLabel + ";};" +
@@ -389,7 +389,7 @@ public class NegationIT {
                     .collect(toList());
             assertCollectionsNonTriviallyEqual(
                     expectedAnswers,
-                    answersWithoutSpecifcTypeAndValue
+                    answersWithoutSpecificTypeAndValue
             );
         }
     }
