+++ test-integration/graql/reasoner/resources/resourceApplicabilityTest.gql
@@ -1,12 +1,12 @@
 define
 
-entity1 sub entity,
+someEntity sub entity,
     has res-string,
     has res-double,
     has res-long,
     has res-boolean,
     has resource;
-entity2 sub entity,
+anotherEntity sub entity,
     has resource;
 
 #Resources
@@ -20,7 +20,7 @@ res-boolean sub attribute, datatype boolean;
 
 rule-1 sub rule,
 when {
-    $x isa entity1;
+    $x isa someEntity;
 },
 then {
     $x has res-string 'string';
@@ -28,7 +28,7 @@ then {
 
 rule-2 sub rule,
 when {
-    $x isa entity1;
+    $x isa someEntity;
 },
 then {
     $x has res-double 3.14;
@@ -36,7 +36,7 @@ then {
 
 rule-3 sub rule,
 when {
-    $x isa entity1;
+    $x isa someEntity;
 },
 then {
     $x has res-long 123;
@@ -44,7 +44,7 @@ then {
 
 rule-4 sub rule,
 when {
-    $x isa entity1;
+    $x isa someEntity;
 },
 then {
     $x has res-boolean 'true';
@@ -52,8 +52,11 @@ then {
 
 rule-5 sub rule,
 when {
-    $x isa entity1;
+    $x isa someEntity;
 },
 then {
     $x has resource 'test';
 };
+
+insert
+$x isa someEntity;
